manupulating_images/requirements.txt
Pillow

mypy
black